Created new iteration of Latest
This new iteration of latest added 4 new autos
- Drive up shoot
- Algae take off
- L3 Auto
- L2

All the autos are stored within AutonomousPeriodic in Robot.java
I did not make them into a new file.

In order to run them
1) Comment out the autos that you're not using
2) Go to driver Station
3) Change the tab from Teleop to Auto
4) Press enable, it should just run
5) If u ever run into an issue, immediately press disable.

The structure of the auto is simple.  There's an if statement that runs based on time (We will do encoders, this is just testing) If the timer is within those parameters it will run.  Otherwise that mechanic won't